Add github actions to test against a matrix of bazel versions
Brought up in the discussion of #463, was that testing against multiple
bazel versions with bazelci didn't seem possible without a fair bit of
work.

Judging based on usage that I have seen, we probably moved too quickly
on bumping the minimum version to 3.0.0, but until we can actually test
against our minimum version, we won't be able to reasonably version
breaking changes as suggested by #415.

This initial PR leaves a lot of on the table in terms of potential
optimizations, and is fixed on Ubuntu 20.04 LTS for the moment.