Add github action to generate and push docs to docs branch
Instead of requiring that users regenerate auto-generated documentation
each time it changes, this adds a github action to force push any
changes on the main branch to the docs branch.  Github pages allows
docs to be on any branch, in the docs folder.

There may be some permissions issues with the docs branch.  In that
case, I'm fine reverting this immediately afterwards.  I've tested this
to the extent I could on my forked repository.